Add assertion
The bug fixed with e9ea6c7 resulted in an argument being a scheme null
pointer. The addtional assertion might help figuring out possible
problems. The assertion in the next line will be thrown anyway if a null
pointer occurs. However, the scheme null pointer is a sentinel object,
so you might not spot it quickly in a debugger session.