Fix tracing
In order to debug the TinyScheme interpreter, you can enable tracing by
adding the following line in your Scheme code:

    (tracing 1)

Unfortunately, this results in crashes that look like memory handling
issues in TinyScheme itself. I have not investigated the crashes further
and simply added a workaround, which can be compiled in by passing
-DSBPL_DEBUG=1 to cmake. The trace will then be simply written to
stderr. This is probably undesired for a general purpose interpreter,
however, it is sufficient in our use case.